refactor(Draw): sort
