+++ pvzclass/include/Draw.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "PVZ.h"
+#include "../PVZ.h"
 
 namespace PVZ
 {
