+++ pvzclass/pvzclass.h
@@ -5,6 +5,6 @@
 #include "Classes.hpp"
 #include "Creators.h"
 #include "Injectors.h"
-#include "Draw.h"
+#include "include/Draw.h"
 #include "Extensions.h"
 #include "include/Sexy.h"
