+++ pvzclass/pvzclass.vcxproj
@@ -194,7 +194,7 @@
     <ClCompile Include="Classes\TrackInstance.cpp" />
     <ClCompile Include="Classes\ZenGarden.cpp" />
     <ClCompile Include="Creators.cpp" />
-    <ClCompile Include="Draw.cpp" />
+    <ClCompile Include="src\Draw.cpp" />
     <ClCompile Include="Enums\CaptionStyle.cpp" />
     <ClCompile Include="Enums\SeedType.cpp" />
     <ClCompile Include="Enums\ChallengePage.cpp" />
@@ -269,7 +269,7 @@
     <ClInclude Include="Classes\TodParticleSystem.hpp" />
     <ClInclude Include="Classes\ZenGarden.hpp" />
     <ClInclude Include="deprecated.hpp" />
-    <ClInclude Include="Draw.h" />
+    <ClInclude Include="src\Draw.h" />
     <ClInclude Include="Enums\KeyCode.hpp" />
     <ClInclude Include="Enums\StoreItem.hpp" />
     <ClInclude Include="Enums\ZombieAttackType.hpp" />
